# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "SchellingBoard API", "version": "1" },
...
  "paths": {
...
    "/api/v1/meetings": {
      "get": {
        "tags": ["meetings"],
        "description": "The acting guest's own 1-on-1s at an event, and the slots they declared themselves free.",
        "parameters": [
          {
            "schema": { "type": "string", "minLength": 1 },
            "required": true,
            "name": "eventId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "The caller's 1-on-1s and availability",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "meetings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "status": {
                            "type": "string",
                            "enum": [
                              "pending",
                              "accepted",
                              "declined",
                              "canceled",
                              "expired"
                            ],
                            "description": "expired: a request nobody answered before its slot began"
                          },
                          "role": {
                            "type": "string",
                            "enum": ["requester", "recipient"]
                          },
                          "otherId": { "type": "string" },
                          "otherName": { "type": "string" },
                          "slotStart": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "slotEnd": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "dayLabel": { "type": "string" },
                          "timeLabel": { "type": "string" },
                          "meetingPoint": { "type": "string" },
                          "message": { "type": "string" },
                          "cancelNote": { "type": "string" },
                          "clashes": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "guestName": { "type": "string" },
                                "kind": {
                                  "type": "string",
                                  "enum": [
                                    "hosting",
                                    "attending",
                                    "meeting",
                                    "busy"
                                  ]
                                },
                                "title": { "type": ["string", "null"] },
                                "isViewer": { "type": "boolean" }
                              },
                              "required": [
                                "guestName",
                                "kind",
                                "title",
                                "isViewer"
                              ],
                              "description": "Only the caller's own commitment is named; another's is busy"
                            }
                          }
                        },
                        "required": [
                          "id",
                          "status",
                          "role",
                          "otherId",
                          "otherName",
                          "slotStart",
                          "slotEnd",
                          "dayLabel",
                          "timeLabel",
                          "meetingPoint",
                          "message",
                          "cancelNote",
                          "clashes"
                        ]
                      },
                      "description": "The caller's own, by slot"
                    },
                    "availability": {
                      "type": "array",
                      "items": { "type": "string", "format": "date-time" },
                      "description": "The slot starts the caller declared themselves free"
                    }
                  },
                  "required": ["meetings", "availability"]
                }
              }
            }
          },
          "default": {
            "description": "Problem details (RFC 9457); clients branch on `code`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": { "type": "string" },
                    "title": { "type": "string" },
                    "status": { "type": "integer" },
                    "code": { "type": "string" },
                    "detail": { "type": "string" },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": { "type": "string" },
                          "message": { "type": "string" }
                        },
                        "required": ["path", "message"]
                      }
                    }
                  },
                  "required": ["type", "title", "status", "code"]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["meetings"],
        "description": "Asks another attendee for a 1-on-1 as the acting guest; the recipient is notified.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Retries with the same key get the first response (24 h)"
            },
            "required": false,
            "description": "Retries with the same key get the first response (24 h)",
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "eventId": { "type": "string" },
                  "recipientId": { "type": "string" },
                  "slotStart": { "type": "string" },
                  "slotCount": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1
                  },
                  "meetingPoint": { "type": "string", "maxLength": 200 },
                  "message": { "type": "string", "maxLength": 2000 }
                },
                "required": [
                  "eventId",
                  "recipientId",
                  "slotStart",
                  "meetingPoint"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The pending request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "eventId": { "type": "string" },
                    "requesterId": { "type": "string" },
                    "recipientId": { "type": "string" },
                    "slotStart": { "type": "string", "format": "date-time" },
                    "slotEnd": { "type": "string", "format": "date-time" },
                    "meetingPoint": { "type": "string" },
                    "message": { "type": "string" },
                    "cancelNote": { "type": "string" },
                    "status": {
                      "type": "string",
                      "enum": ["pending", "accepted", "declined", "canceled"]
                    },
                    "createdAt": { "type": "string", "format": "date-time" },
                    "respondedAt": {
                      "type": ["string", "null"],
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "eventId",
                    "requesterId",
                    "recipientId",
                    "slotStart",
                    "slotEnd",
                    "meetingPoint",
                    "message",
                    "cancelNote",
                    "status",
                    "createdAt",
                    "respondedAt"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Problem details (RFC 9457); clients branch on `code`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": { "type": "string" },
                    "title": { "type": "string" },
                    "status": { "type": "integer" },
                    "code": { "type": "string" },
                    "detail": { "type": "string" },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": { "type": "string" },
                          "message": { "type": "string" }
                        },
                        "required": ["path", "message"]
                      }
                    }
                  },
                  "required": ["type", "title", "status", "code"]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}